Using uintmax_t so the int is big enough for the type enif_get_uint64 expects

+++ c_src/enacl_nif.c
@@ -837,7 +837,7 @@ ERL_NIF_TERM enif_crypto_onetimeauth_verify(ErlNifEnv *env, int argc, ERL_NIF_TE
 static
 ERL_NIF_TERM enif_randombytes(ErlNifEnv *env, int argc, ERL_NIF_TERM const argv[])
 {
-	size_t req_size;
+	uintmax_t req_size;
 	ErlNifBinary result;
 
 	if ((argc != 1) || (!enif_get_uint64(env, argv[0], &req_size))) {
